FaceAPI now effectively detects a face in an image selected from device's local gallery.
I have returned the Vuzix action menu to the app and thus replaced the button with a menu item. There are currently two menu items: one for browsing the device's local gallery (which works), and one for using the camera (which only partially works; camera usage works, but detection won't run).

TODO:
- PHASE 1: use the sample code from Microsoft to implement actual recognition, with training
- CONTINUOUS: organize the code better (use Microsoft's sample code)
- PHASE 1: enable using picture taken by the camera for detection & recognition
- PHASE 2: create a "live mode"